pyproject.toml
[project]
name = "agent-analyser"
version = "0.1.0"
description = "Copilot Studio bot export analyser — parses botContent.yml and dialog.json to generate Markdown reports with Mermaid sequence diagrams"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "reflex>=0.8.27",
    "typer>=0.23.1",
]

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.1",
]
